+++ scripts/clean-modules/index.js
@@ -1,4 +1,0 @@
-/* eslint-disable import/no-extraneous-dependencies */
-const shell = require('shelljs');
-
-shell.rm('-rf', 'node_modules');